# Enable full LTO on Python 3.12 and 3.13