Closes https://github.com/astral-sh/python-build-standalone/issues/528