@zanieb 2025-02-13
I'll try to do a brief analysis of the performance and build time trade-offs here.