@zanieb 2025-02-13
Surprisingly, this is consistently a bit slower on the pystones benchmark used for #524 

```
❯ hyperfine "$(uv python find 3.12.9) ./pystones.py" "./python/install/bin/python ./pystones.py"
Benchmark 1: /Users/zb/.local/share/uv/python/cpython-3.12.9-macos-aarch64-none/bin/python3.12 ./pystones.py
  Time (mean ± σ):      75.1 ms ±   1.0 ms    [User: 72.3 ms, System: 2.1 ms]
  Range (min … max):    73.2 ms …  76.6 ms    39 runs
 
Benchmark 2: ./python/install/bin/python ./pystones.py
  Time (mean ± σ):      79.2 ms ±   1.0 ms    [User: 76.0 ms, System: 2.5 ms]
  Range (min … max):    77.6 ms …  81.7 ms    36 runs
 
Summary
  /Users/zb/.local/share/uv/python/cpython-3.12.9-macos-aarch64-none/bin/python3.12 ./pystones.py ran
    1.05 ± 0.02 times faster than ./python/install/bin/python ./pystones.py
```