@zanieb 2025-02-13
This is less pronounced (but still present?) if I compare to an artifact from `main` instead of a uv-installed binary

```
❯ hyperfine "./main/install/bin/python3.13 ./pystones.py" "./branch/install/bin/python3.13 ./pystones.py" --min-runs 100
Benchmark 1: ./main/install/bin/python3.13 ./pystones.py
  Time (mean ± σ):      77.4 ms ±   1.1 ms    [User: 74.5 ms, System: 2.3 ms]
  Range (min … max):    75.1 ms …  80.4 ms    100 runs
 
Benchmark 2: ./branch/install/bin/python3.13 ./pystones.py
  Time (mean ± σ):      79.2 ms ±   1.1 ms    [User: 76.3 ms, System: 2.3 ms]
  Range (min … max):    77.4 ms …  85.0 ms    100 runs
 
Summary
  ./main/install/bin/python3.13 ./pystones.py ran
    1.02 ± 0.02 times faster than ./branch/install/bin/python3.13 ./pystones.py
```