@zanieb 2025-02-13
I'll find another benchmark to try.

The summary at https://github.com/python/cpython/issues/122580#issuecomment-2325733458 is relatively compelling.