@zanieb 2025-02-13
With the script from https://github.com/python/cpython/issues/122580#issue-2443289218 — which was the original reproduction for a regression, I also don't see an improvement

```
❯ hyperfine "./main/install/bin/python3.13 ./repro.py" "./branch/install/bin/python3.13 ./repro.py"
Benchmark 1: ./main/install/bin/python3.13 ./repro.py
  Time (mean ± σ):      5.926 s ±  0.048 s    [User: 5.900 s, System: 0.021 s]
  Range (min … max):    5.833 s …  5.975 s    10 runs
 
Benchmark 2: ./branch/install/bin/python3.13 ./repro.py
  Time (mean ± σ):      5.988 s ±  0.035 s    [User: 5.962 s, System: 0.018 s]
  Range (min … max):    5.927 s …  6.050 s    10 runs
 
Summary
  ./main/install/bin/python3.13 ./repro.py ran
    1.01 ± 0.01 times faster than ./branch/install/bin/python3.13 ./repro.py
```