@Fidget-Spinner 2025-02-13
@zanieb It might be that this was a bug in Apple Clang but fixed in normal Clang. The version of clang Ned used was `clang-1500` but knowing Apple, that isn't actually clang 15 but some older internal version. The Clang version here is pretty new in comparison.

So I'm sorry for the noise if that turns out to be the case.