@Fidget-Spinner 2025-02-13
I think my guess is probably right. I will close the issue as the build time tradeoff is not worth it. So sorry for the noise!