@zanieb 2025-02-25
I'm going to explore this again in the context of #535 and #539 — I think thin LTO is one of the few configuration differences between our build and the conda-forge one.

I'll test some of the binaries here with the pyperformance suite.